Changed text and import codes
I changed the Design Heritage text that was copied to the Visual Communication text that was needed for this pages. It wouldn't import properly so i had to check as to what was wrong. I noticed that the Header text was not changed to Header2 so i changed this and it looked how it was intended to look.